+++ src/LaravelFacebookSdk/LaravelUrlDetectionHandler.php
@@ -1,16 +1,28 @@
 <?php namespace SammyK\LaravelFacebookSdk;
 
+use Illuminate\Routing\UrlGenerator;
 use Facebook\Url\UrlDetectionInterface;
 
 class LaravelUrlDetectionHandler implements UrlDetectionInterface
 {
+    /**
+     * @var UrlGenerator
+     */
+    private $url;
+
+    /**
+     * @param UrlGenerator $url
+     */
+    public function __construct(UrlGenerator $url)
+    {
+        $this->url = $url;
+    }
 
     /**
      * @inheritdoc
      */
     public function getCurrentUrl()
     {
-        return \Request::url();
+        return $this->url->current();
     }
-
 }