+++ src/views/channel.blade.php
@@ -1,1 +1,0 @@
-<script src="//connect.facebook.net/{{ Config::get('laravel-facebook-sdk::locale') }}/all.js"></script>