+++ tests/LaravelFacebookSdkTest.php
@@ -1,15 +1,20 @@
-<?php
+<?php namespace SammyK\LaravelFacebookSdk\Test;
 
 use Mockery as m;
 use SammyK\LaravelFacebookSdk\LaravelFacebookSdk;
 
-class LaravelFacebookSdkTest extends PHPUnit_Framework_TestCase
+class LaravelFacebookSdkTest extends \PHPUnit_Framework_TestCase
 {
 
     /**
-     * @var \Illuminate\Config\Repository
+     * @var \Illuminate\Config\Repository|\Mockery\MockInterface
      */
     protected $config_mock;
+
+    /**
+     * @var \Illuminate\Routing\UrlGenerator|\Mockery\MockInterface
+     */
+    protected $url_mock;
 
     /**
      * @var LaravelFacebookSdk
@@ -19,7 +24,8 @@ class LaravelFacebookSdkTest extends PHPUnit_Framework_TestCase
     public function setUp()
     {
         $this->config_mock = m::mock('Illuminate\Config\Repository');
-        $this->laravel_facebook_sdk = new LaravelFacebookSdk($this->config_mock, [
+        $this->url_mock = m::mock('Illuminate\Routing\UrlGenerator');
+        $this->laravel_facebook_sdk = new LaravelFacebookSdk($this->config_mock, $this->url_mock, [
             'app_id' => 'foo_id',
             'app_secret' => 'foo_secret',
             'persistent_data_handler' => 'memory',
@@ -34,19 +40,20 @@ class LaravelFacebookSdkTest extends PHPUnit_Framework_TestCase
     /** @test */
     public function when_no_arguments_are_passed_the_get_login_url_method_will_default_to_config()
     {
+        $this->url_mock
+            ->shouldReceive('to')
+            ->with('/foo')
+            ->once()
+            ->andReturn('https://foohost/foo');
+
         $this->config_mock
             ->shouldReceive('get')
-            ->with('laravel-facebook-sdk::default_scope')
+            ->with('laravel-facebook-sdk.default_scope')
             ->once()
             ->andReturn(['foo', 'bar']);
         $this->config_mock
             ->shouldReceive('get')
-            ->with('app.url')
-            ->once()
-            ->andReturn('https://foohost');
-        $this->config_mock
-            ->shouldReceive('get')
-            ->with('laravel-facebook-sdk::default_redirect_uri')
+            ->with('laravel-facebook-sdk.default_redirect_uri')
             ->once()
             ->andReturn('/foo');
 
@@ -59,6 +66,12 @@ class LaravelFacebookSdkTest extends PHPUnit_Framework_TestCase
     /** @test */
     public function the_default_config_can_be_overwritten_by_passing_arguments_to_get_login_url()
     {
+        $this->url_mock
+            ->shouldReceive('to')
+            ->with('https://poop.fart/callback')
+            ->once()
+            ->andReturn('https://poop.fart/callback');
+
         $this->config_mock
             ->shouldReceive('get')
             ->never();